requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cffi==1.15.1
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
py-cord==2.4.1
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
typing-extensions==4.5.0
yarl==1.8.2
